feat: support important and negative modifiers

+++ src/plugins/layout.rs
@@ -240,7 +240,7 @@ impl Plugin for LayoutZIndexPlugin {
 
     fn get_css_for_modifier(&self, modifier: &str) -> Option<String> {
         // NOTE: Not-compatible with TailwindCSS, support all values
-        if modifier.parse::<usize>().is_ok() || is_matching_auto(modifier) {
+        if modifier.parse::<f32>().is_ok() || is_matching_auto(modifier) {
             Some(format!("z-index: {modifier};"))
         } else {
             None
